Extract AsyncProfilerTask
